[MINOR] update readme on PIN-SLAM with colorized lidar points

+++ dataset/dataloaders/kitti360.py
@@ -191,8 +191,8 @@ class KITTI360Dataset:
         points_proj = np.matmul(K_mat[:3,:3].reshape([1,3,3]), points)
         depth = points_proj[:,2,:]
         depth[depth==0] = -1e-6
-        u = np.round(points_proj[:,0,:]/np.abs(depth)).astype(np.int32)
-        v = np.round(points_proj[:,1,:]/np.abs(depth)).astype(np.int32)
+        u = np.round(points_proj[:,0,:]/np.abs(depth)).astype(int)
+        v = np.round(points_proj[:,1,:]/np.abs(depth)).astype(int)
 
         if ndim==2:
             u = u[0]; v=v[0]; depth=depth[0]
